# docs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduMark",
    "description": "\n基于大模型的中小学作业批改系统，提供以下核心能力：\n\n- **作业提交与批改**：上传作业图片，自动识别题目并进行 AI 批改\n- **知识库管理**：上传教学资料，构建知识库辅助批改\n- **用户管理**：学生/老师/管理员三种角色，权限分级\n- **模型管理**：配置视觉模型和语言模型\n- **配置管理**：个性化识别模式与增强选项\n",
    "contact": {
      "name": "EduMark Team",
      "email": "admin@edumark.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/knowledge/{id}": {
      "get": {
        "tags": [
          "知识文档"
        ],
        "summary": "查询知识文档",
        "description": "根据 ID 查询单个知识文档的详细信息。",
        "operationId": "get_knowledge_api_v1_knowledge__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "知识文档 ID",
              "title": "Id"
            },
            "description": "知识文档 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功，返回文档信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_KnowledgeOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "知识文档"
        ],
        "summary": "删除知识文档",
        "description": "删除知识文档，同步清理关联数据。\n\n- 同时删除 MinIO 中的文件\n- 同时删除 Qdrant 中的向量数据\n- 只能删除自己的文档，管理员可删除任意文档",
        "operationId": "delete_knowledge_api_v1_knowledge__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "知识文档 ID",
              "title": "Id"
            },
            "description": "知识文档 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "403": {
            "description": "无权删除该文档"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/page": {
      "post": {
        "tags": [
          "知识文档"
        ],
        "summary": "分页查询知识文档",
        "description": "分页查询知识文档列表。\n\n- 支持排序和过滤条件\n- 默认每页 10 条，最大 100 条",
        "operationId": "page_knowledge_api_v1_knowledge_page_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageDTO"
              },
              "examples": {
                "默认分页": {
                  "summary": "默认分页查询（第1页，每页10条）",
                  "value": {
                    "page_num": 1,
                    "page_size": 10
                  }
                },
                "按科目过滤": {
                  "summary": "筛选数学科目的知识文档",
                  "value": {
                    "page_num": 1,
                    "page_size": 10,
                    "query": {
                      "subject": "数学"
                    }
                  }
                },
                "按年级过滤": {
                  "summary": "筛选二年级的知识文档",
                  "value": {
                    "page_num": 1,
                    "page_size": 10,
                    "query": {
                      "grade": "二年级"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "查询成功，返回分页数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PageVO_KnowledgeOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/upload": {
      "post": {
        "tags": [
          "知识文档"
        ],
        "summary": "上传知识文档",
        "description": "上传知识文档文件，异步解析入库。\n\n- 支持上传 PDF、Word 等文档格式\n- 上传后自动异步解析，状态由 `pending` → `parsing` → `done`\n- 解析完成后内容存入向量数据库，供批改时检索\n- 仅老师和管理员可上传",
        "operationId": "upload_knowledge_api_v1_knowledge_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_knowledge_api_v1_knowledge_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "上传成功，返回文档信息，解析将在后台异步进行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_KnowledgeOut_"
                }
              }
            }
          },
          "403": {
            "description": "仅老师和管理员可上传"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/models/{id}": {
      "get": {
        "tags": [
          "模型管理"
        ],
        "summary": "查询模型",
        "description": "根据 ID 查询单个模型信息。",
        "operationId": "get_model_api_v1_models__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "模型 ID",
              "title": "Id"
            },
            "description": "模型 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功，返回模型信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_ModelOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "模型管理"
        ],
        "summary": "更新模型",
        "description": "更新模型信息，仅管理员可操作。",
        "operationId": "update_model_api_v1_models__id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "模型 ID",
              "title": "Id"
            },
            "description": "模型 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              },
              "examples": {
                "更新视觉模型": {
                  "summary": "更新为新的视觉模型名称",
                  "value": {
                    "name": "qwen-vl-plus",
                    "mode": 1
                  }
                },
                "更新语言模型": {
                  "summary": "更新为新的语言模型名称",
                  "value": {
                    "name": "qwen-plus",
                    "mode": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功，返回模型信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_ModelOut_"
                }
              }
            }
          },
          "403": {
            "description": "仅管理员可操作"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "模型管理"
        ],
        "summary": "删除模型",
        "description": "删除模型配置，仅管理员可操作。",
        "operationId": "delete_model_api_v1_models__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "模型 ID",
              "title": "Id"
            },
            "description": "模型 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "403": {
            "description": "仅管理员可操作"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models": {
      "post": {
        "tags": [
          "模型管理"
        ],
        "summary": "新增模型",
        "description": "添加新的 AI 模型配置，仅管理员可操作。\n\n- **mode=1**：视觉模型，用于图片识别\n- **mode=2**：语言模型，用于文本批改",
        "operationId": "create_model_api_v1_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              },
              "examples": {
                "新增视觉模型": {
                  "summary": "添加一个视觉模型（mode=1）",
                  "value": {
                    "name": "qwen-vl-max",
                    "mode": 1
                  }
                },
                "新增语言模型": {
                  "summary": "添加一个语言模型（mode=2）",
                  "value": {
                    "name": "qwen-max",
                    "mode": 2
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建成功，返回模型信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_ModelOut_"
                }
              }
            }
          },
          "403": {
            "description": "仅管理员可操作"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/models/page": {
      "post": {
        "tags": [
          "模型管理"
        ],
        "summary": "分页查询模型",
        "description": "分页查询模型列表。\n\n- 支持排序和过滤条件\n- 默认每页 10 条，最大 100 条",
        "operationId": "page_model_api_v1_models_page_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageDTO"
              },
              "examples": {
                "默认分页": {
                  "summary": "默认分页查询（第1页，每页10条）",
                  "value": {
                    "page_num": 1,
                    "page_size": 10
                  }
                },
                "按类型过滤": {
                  "summary": "仅查看视觉模型（mode=1）",
                  "value": {
                    "page_num": 1,
                    "page_size": 10,
                    "query": {
                      "mode": 1
                    }
                  }
                },
                "按名称排序": {
                  "summary": "按模型名称升序排列",
                  "value": {
                    "page_num": 1,
                    "page_size": 10,
                    "sort_fields": [
                      {
                        "field": "name",
                        "direction": "asc"
                      }
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "查询成功，返回分页数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_PageVO_ModelOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_knowledge_api_v1_knowledge_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "知识文档文件"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject",
            "description": "科目，如语文、数学"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade",
            "description": "年级，如一年级、二年级"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_knowledge_api_v1_knowledge_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "知识文档 ID"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "所属用户 ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "知识文档标题"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "知识文档存储路径"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject",
            "description": "科目"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade",
            "description": "年级"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "文档状态"
          },
          "chunk": {
            "type": "integer",
            "title": "Chunk",
            "description": "解析后分块数量"
          },
          "create_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Create Time",
            "description": "创建时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "url",
          "status",
          "chunk"
        ],
        "title": "KnowledgeOut",
        "example": {
          "chunk": 15,
          "create_time": "2026-06-04T09:00:00",
          "grade": "二年级",
          "id": 1,
          "status": "completed",
          "subject": "数学",
          "title": "二年级数学知识点汇总.pdf",
          "url": "knowledge/2/1717488000_math_knowledge.pdf",
          "user_id": 2
        }
      },
...
      "ModelCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "description": "模型名称，1~64 个字符"
          },
          "mode": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Mode",
            "description": "模型类型：1=视觉模型，2=语言模型"
          }
        },
        "type": "object",
        "required": [
          "name",
          "mode"
        ],
        "title": "ModelCreate"
      },
      "ModelOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "模型记录 ID"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "模型名称"
          },
          "mode": {
            "type": "integer",
            "title": "Mode",
            "description": "模型类型：1=视觉模型，2=语言模型"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "mode"
        ],
        "title": "ModelOut",
        "example": {
          "id": 1,
          "mode": 1,
          "name": "qwen-vl-max"
        }
      },
      "PageDTO": {
        "properties": {
          "page_num": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Num",
            "description": "页码，从 1 开始",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "description": "每页条数，范围 1~100",
            "default": 10
          },
          "sort_fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SortField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Fields",
            "description": "排序规则列表，可选"
          },
          "query": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "查询过滤条件，键值对形式，可选"
          }
        },
        "type": "object",
        "title": "PageDTO"
      },
      "PageVO_KnowledgeOut_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "总记录数"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeOut"
            },
            "type": "array",
            "title": "Rows",
            "description": "当前页数据列表"
          }
        },
        "type": "object",
        "required": [
          "total",
          "rows"
        ],
        "title": "PageVO[KnowledgeOut]",
        "example": {
          "rows": [],
          "total": 100
        }
      },
...
      "Result": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "Result_KnowledgeOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[KnowledgeOut]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
      "Result_ModelOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[ModelOut]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
      "Result_PageVO_KnowledgeOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageVO_KnowledgeOut_"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[PageVO[KnowledgeOut]]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
      "Result_PageVO_ModelOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageVO_ModelOut_"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[PageVO[ModelOut]]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "SortField": {
        "properties": {
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field",
            "description": "排序字段名"
          },
          "direction": {
            "$ref": "#/components/schemas/SortDirection",
            "description": "排序方向，asc=升序，desc=降序",
            "default": "asc"
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "SortField"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}